refactor(api): split TrainerRelationshipController into focused partials
Extract invitations-specific logic into TrainerRelationshipController.Invitations partial.
Reduces main controller to 60 lines, improving maintainability per 300-line rule.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
