+++ LgymApi.Api/Features/Trainer/Controllers/TrainerRelationshipController.cs
@@ -31,7 +31,7 @@ namespace LgymApi.Api.Features.Trainer.Controllers;
 [ApiController]
 [Route("api/trainer")]
 [Authorize(Policy = AuthConstants.Policies.TrainerAccess)]
-public sealed class TrainerRelationshipController : ControllerBase
+public sealed partial class TrainerRelationshipController : ControllerBase
 {
     private readonly ITrainerRelationshipService _trainerRelationshipService;
     private readonly IMapper _mapper;
@@ -40,380 +40,5 @@ public sealed class TrainerRelationshipController : ControllerBase
     {
         _trainerRelationshipService = trainerRelationshipService;
         _mapper = mapper;
-    }
-
-    [HttpPost("invitations")]
-    [ApiIdempotency("/api/trainer/invitations", ApiIdempotencyScopeSource.AuthenticatedUser)]
-    [ProducesResponseType(typeof(TrainerInvitationDto), StatusCodes.Status200OK)]
-    public async Task<IActionResult> CreateInvitation([FromBody] CreateTrainerInvitationRequest request, CancellationToken cancellationToken = default)
-    {
-        if (!Id<UserEntity>.TryParse(request.TraineeId, out var traineeId))
-        {
-            return Result<TrainerInvitationResult, AppError>.Failure(new InvalidTrainerRelationshipError(Messages.UserIdRequired)).ToActionResult();
-        }
-
-        var trainer = HttpContext.GetCurrentUser();
-        var result = await _trainerRelationshipService.CreateInvitationAsync(trainer!, traineeId, cancellationToken);
-        if (result.IsFailure)
-        {
-            return result.ToActionResult();
-        }
-
-        return Ok(_mapper.Map<LgymApi.Application.Features.TrainerRelationships.Models.TrainerInvitationResult, TrainerInvitationDto>(result.Value));
-    }
-
-    [HttpPost("invitations/paginated")]
-    [ProducesResponseType(typeof(PaginatedTrainerInvitationResult), StatusCodes.Status200OK)]
-    public async Task<IActionResult> GetInvitationsPaginated([FromBody] PaginatedTrainerInvitationRequest request, CancellationToken cancellationToken = default)
-    {
-        var filterInput = new FilterInput
-        {
-            Page = request.Page,
-            PageSize = request.PageSize,
-            FilterGroups = request.FilterGroups,
-            SortDescriptors = request.SortDescriptors
-        };
-
-        var trainer = HttpContext.GetCurrentUser();
-        var result = await _trainerRelationshipService.GetInvitationsPaginatedAsync(trainer!, filterInput, cancellationToken);
-        if (result.IsFailure)
-        {
-            return result.ToActionResult();
-        }
-
-        var pagination = result.Value;
-        var response = new PaginatedTrainerInvitationResult
-        {
-            Items = _mapper.MapList<TrainerInvitationResult, TrainerInvitationDto>(pagination.Items),
-            Page = pagination.Page,
-            PageSize = pagination.PageSize,
-            TotalCount = pagination.TotalCount,
-            TotalPages = pagination.TotalPages,
-            HasNextPage = pagination.HasNextPage,
-            HasPreviousPage = pagination.HasPreviousPage
-        };
-        return Ok(response);
-    }
-
-    [HttpPost("invitations/by-email")]
-    [ProducesResponseType(typeof(TrainerInvitationDto), StatusCodes.Status200OK)]
-    public async Task<IActionResult> CreateInvitationByEmail([FromBody] CreateTrainerInvitationByEmailRequest request, CancellationToken cancellationToken = default)
-    {
-        var trainer = HttpContext.GetCurrentUser();
-        var result = await _trainerRelationshipService.CreateInvitationByEmailAsync(
-            trainer!, request.Email, request.PreferredLanguage, request.PreferredTimeZone, cancellationToken);
-        if (result.IsFailure)
-        {
-            return result.ToActionResult();
-        }
-
-        return Ok(_mapper.Map<LgymApi.Application.Features.TrainerRelationships.Models.TrainerInvitationResult, TrainerInvitationDto>(result.Value));
-    }
-
-    [HttpPost("invitations/{invitationId}/revoke")]
-    [ProducesResponseType(typeof(ResponseMessageDto), StatusCodes.Status200OK)]
-    public async Task<IActionResult> RevokeInvitation([FromRoute] string invitationId, CancellationToken cancellationToken = default)
-    {
-        if (!Id<TrainerInvitationEntity>.TryParse(invitationId, out var parsedInvitationId))
-        {
-            return Result<Unit, AppError>.Failure(new InvalidTrainerRelationshipError(Messages.FieldRequired)).ToActionResult();
-        }
-
-        var trainer = HttpContext.GetCurrentUser();
-        var result = await _trainerRelationshipService.RevokeInvitationAsync(trainer!, parsedInvitationId, cancellationToken);
-        if (result.IsFailure)
-        {
-            return result.ToActionResult();
-        }
-
-        return Ok(_mapper.Map<string, ResponseMessageDto>(Messages.Updated));
-    }
-
-    [HttpGet("trainees")]
-    [ProducesResponseType(typeof(TrainerDashboardTraineesResponse), StatusCodes.Status200OK)]
-    public async Task<IActionResult> GetDashboardTrainees([FromQuery] TrainerDashboardTraineesRequest request, CancellationToken cancellationToken = default)
-    {
-        var trainer = HttpContext.GetCurrentUser();
-        var result = await _trainerRelationshipService.GetDashboardTraineesAsync(trainer!, new TrainerDashboardTraineeQuery
-        {
-            Search = request.Search,
-            Status = request.Status,
-            SortBy = request.SortBy,
-            SortDirection = request.SortDirection,
-            Page = request.Page,
-            PageSize = request.PageSize
-        }, cancellationToken);
-
-        if (result.IsFailure)
-        {
-            return result.ToActionResult();
-        }
-
-        return Ok(new TrainerDashboardTraineesResponse
-        {
-            Page = result.Value.Page,
-            PageSize = result.Value.PageSize,
-            Total = result.Value.Total,
-            Items = _mapper.MapList<TrainerDashboardTraineeResult, TrainerDashboardTraineeDto>(result.Value.Items)
-        });
-    }
-
-    [HttpGet("trainees/{traineeId}/trainings/dates")]
-    [ProducesResponseType(typeof(List<DateTime>), StatusCodes.Status200OK)]
-    [ProducesResponseType(typeof(ResponseMessageDto), StatusCodes.Status400BadRequest)]
-    [ProducesResponseType(typeof(ResponseMessageDto), StatusCodes.Status404NotFound)]
-    public async Task<IActionResult> GetTraineeTrainingDates([FromRoute] string traineeId, CancellationToken cancellationToken = default)
-    {
-        if (!Id<UserEntity>.TryParse(traineeId, out var parsedTraineeId))
-        {
-            return Result<List<DateTime>, AppError>.Failure(new InvalidTrainerRelationshipError(Messages.UserIdRequired)).ToActionResult();
-        }
-
-        var trainer = HttpContext.GetCurrentUser();
-        var result = await _trainerRelationshipService.GetTraineeTrainingDatesAsync(trainer!, parsedTraineeId, cancellationToken);
-        if (result.IsFailure)
-        {
-            return result.ToActionResult();
-        }
-
-        return Ok(result.Value);
-    }
-
-    [HttpPost("trainees/{traineeId}/trainings/by-date")]
-    [ProducesResponseType(typeof(List<TrainingByDateDetailsDto>), StatusCodes.Status200OK)]
-    [ProducesResponseType(typeof(ResponseMessageDto), StatusCodes.Status400BadRequest)]
-    [ProducesResponseType(typeof(ResponseMessageDto), StatusCodes.Status404NotFound)]
-    public async Task<IActionResult> GetTraineeTrainingByDate([FromRoute] string traineeId, [FromBody] TrainingByDateRequestDto request, CancellationToken cancellationToken = default)
-    {
-        if (!Id<UserEntity>.TryParse(traineeId, out var parsedTraineeId))
-        {
-            return Result<List<TrainingByDateDetails>, AppError>.Failure(new InvalidTrainerRelationshipError(Messages.UserIdRequired)).ToActionResult();
-        }
-
-        var trainer = HttpContext.GetCurrentUser();
-        var result = await _trainerRelationshipService.GetTraineeTrainingByDateAsync(trainer!, parsedTraineeId, request.CreatedAt, cancellationToken);
-        if (result.IsFailure)
-        {
-            return result.ToActionResult();
-        }
-
-        return Ok(_mapper.MapList<TrainingByDateDetails, TrainingByDateDetailsDto>(result.Value));
-    }
-
-    [HttpPost("trainees/{traineeId}/exercise-scores/chart")]
-    [ProducesResponseType(typeof(List<ExerciseScoresChartDataDto>), StatusCodes.Status200OK)]
-    [ProducesResponseType(typeof(ResponseMessageDto), StatusCodes.Status400BadRequest)]
-    [ProducesResponseType(typeof(ResponseMessageDto), StatusCodes.Status404NotFound)]
-    public async Task<IActionResult> GetTraineeExerciseScoresChartData([FromRoute] string traineeId, [FromBody] ExerciseScoresChartRequestDto request, CancellationToken cancellationToken = default)
-    {
-        if (!Id<UserEntity>.TryParse(traineeId, out var parsedTraineeId))
-        {
-            return Result<List<ExerciseScoresChartData>, AppError>.Failure(new InvalidTrainerRelationshipError(Messages.UserIdRequired)).ToActionResult();
-        }
-
-        if (!Id<ExerciseEntity>.TryParse(request.ExerciseId, out var parsedExerciseId))
-        {
-            return Result<List<ExerciseScoresChartData>, AppError>.Failure(new InvalidTrainerRelationshipError(Messages.ExerciseIdRequired)).ToActionResult();
-        }
-
-        var trainer = HttpContext.GetCurrentUser();
-        var result = await _trainerRelationshipService.GetTraineeExerciseScoresChartDataAsync(trainer!, parsedTraineeId, parsedExerciseId, cancellationToken);
-        if (result.IsFailure)
-        {
-            return result.ToActionResult();
-        }
-
-        return Ok(_mapper.MapList<ExerciseScoresChartData, ExerciseScoresChartDataDto>(result.Value));
-    }
-
-    [HttpGet("trainees/{traineeId}/elo/chart")]
-    [ProducesResponseType(typeof(List<EloRegistryBaseChartDto>), StatusCodes.Status200OK)]
-    [ProducesResponseType(typeof(ResponseMessageDto), StatusCodes.Status400BadRequest)]
-    [ProducesResponseType(typeof(ResponseMessageDto), StatusCodes.Status404NotFound)]
-    public async Task<IActionResult> GetTraineeEloChart([FromRoute] string traineeId, CancellationToken cancellationToken = default)
-    {
-        if (!Id<UserEntity>.TryParse(traineeId, out var parsedTraineeId))
-        {
-            return Result<List<EloRegistryChartEntry>, AppError>.Failure(new InvalidTrainerRelationshipError(Messages.UserIdRequired)).ToActionResult();
-        }
-
-        var trainer = HttpContext.GetCurrentUser();
-        var result = await _trainerRelationshipService.GetTraineeEloChartAsync(trainer!, parsedTraineeId, cancellationToken);
-        if (result.IsFailure)
-        {
-            return result.ToActionResult();
-        }
-
-        return Ok(_mapper.MapList<EloRegistryChartEntry, EloRegistryBaseChartDto>(result.Value));
-    }
-
-    [HttpGet("trainees/{traineeId}/main-records/history")]
-    [ProducesResponseType(typeof(List<MainRecordResponseDto>), StatusCodes.Status200OK)]
-    [ProducesResponseType(typeof(ResponseMessageDto), StatusCodes.Status400BadRequest)]
-    [ProducesResponseType(typeof(ResponseMessageDto), StatusCodes.Status404NotFound)]
-    public async Task<IActionResult> GetTraineeMainRecordsHistory([FromRoute] string traineeId, CancellationToken cancellationToken = default)
-    {
-        if (!Id<UserEntity>.TryParse(traineeId, out var parsedTraineeId))
-        {
-            return Result<List<LgymApi.Domain.Entities.MainRecord>, AppError>.Failure(new InvalidTrainerRelationshipError(Messages.UserIdRequired)).ToActionResult();
-        }
-
-        var trainer = HttpContext.GetCurrentUser();
-        var result = await _trainerRelationshipService.GetTraineeMainRecordsHistoryAsync(trainer!, parsedTraineeId, cancellationToken);
-        if (result.IsFailure)
-        {
-            return result.ToActionResult();
-        }
-
-        var mappedRecords = _mapper.MapList<LgymApi.Domain.Entities.MainRecord, MainRecordResponseDto>(result.Value);
-        return Ok(mappedRecords);
-    }
-
-    [HttpPost("trainees/{traineeId}/unlink")]
-    [ProducesResponseType(typeof(ResponseMessageDto), StatusCodes.Status200OK)]
-    public async Task<IActionResult> UnlinkTrainee([FromRoute] string traineeId, CancellationToken cancellationToken = default)
-    {
-        if (!Id<UserEntity>.TryParse(traineeId, out var parsedTraineeId))
-        {
-            return Result<Unit, AppError>.Failure(new InvalidTrainerRelationshipError(Messages.UserIdRequired)).ToActionResult();
-        }
-
-        var trainer = HttpContext.GetCurrentUser();
-        var result = await _trainerRelationshipService.UnlinkTraineeAsync(trainer!, parsedTraineeId, cancellationToken);
-        if (result.IsFailure)
-        {
-            return result.ToActionResult();
-        }
-
-        return Ok(_mapper.Map<string, ResponseMessageDto>(Messages.Updated));
-    }
-
-    [HttpGet("trainees/{traineeId}/plans")]
-    [ProducesResponseType(typeof(List<TrainerManagedPlanDto>), StatusCodes.Status200OK)]
-    public async Task<IActionResult> GetTraineePlans([FromRoute] string traineeId, CancellationToken cancellationToken = default)
-    {
-        if (!Id<UserEntity>.TryParse(traineeId, out var parsedTraineeId))
-        {
-            return Result<List<TrainerManagedPlanResult>, AppError>.Failure(new InvalidTrainerRelationshipError(Messages.UserIdRequired)).ToActionResult();
-        }
-
-        var trainer = HttpContext.GetCurrentUser();
-        var result = await _trainerRelationshipService.GetTraineePlansAsync(trainer!, parsedTraineeId, cancellationToken);
-        if (result.IsFailure)
-        {
-            return result.ToActionResult();
-        }
-
-        return Ok(_mapper.MapList<TrainerManagedPlanResult, TrainerManagedPlanDto>(result.Value));
-    }
-
-    [HttpPost("trainees/{traineeId}/plans")]
-    [ProducesResponseType(typeof(TrainerManagedPlanDto), StatusCodes.Status201Created)]
-    public async Task<IActionResult> CreateTraineePlan([FromRoute] string traineeId, [FromBody] TrainerPlanFormRequest request, CancellationToken cancellationToken = default)
-    {
-        if (!Id<UserEntity>.TryParse(traineeId, out var parsedTraineeId))
-        {
-            return Result<TrainerManagedPlanResult, AppError>.Failure(new InvalidTrainerRelationshipError(Messages.UserIdRequired)).ToActionResult();
-        }
-
-        var trainer = HttpContext.GetCurrentUser();
-        var result = await _trainerRelationshipService.CreateTraineePlanAsync(trainer!, parsedTraineeId, request.Name, cancellationToken);
-        if (result.IsFailure)
-        {
-            return result.ToActionResult();
-        }
-
-        return StatusCode(StatusCodes.Status201Created, _mapper.Map<TrainerManagedPlanResult, TrainerManagedPlanDto>(result.Value));
-    }
-
-    [HttpPost("trainees/{traineeId}/plans/{planId}/update")]
-    [ProducesResponseType(typeof(TrainerManagedPlanDto), StatusCodes.Status200OK)]
-    public async Task<IActionResult> UpdateTraineePlan([FromRoute] string traineeId, [FromRoute] string planId, [FromBody] TrainerPlanFormRequest request, CancellationToken cancellationToken = default)
-    {
-        if (!Id<UserEntity>.TryParse(traineeId, out var parsedTraineeId))
-        {
-            return Result<TrainerManagedPlanResult, AppError>.Failure(new InvalidTrainerRelationshipError(Messages.UserIdRequired)).ToActionResult();
-        }
-
-        if (!Id<PlanEntity>.TryParse(planId, out var parsedPlanId))
-        {
-            return Result<TrainerManagedPlanResult, AppError>.Failure(new InvalidTrainerRelationshipError(Messages.FieldRequired)).ToActionResult();
-        }
-
-        var trainer = HttpContext.GetCurrentUser();
-        var result = await _trainerRelationshipService.UpdateTraineePlanAsync(trainer!, parsedTraineeId, parsedPlanId, request.Name, cancellationToken);
-        if (result.IsFailure)
-        {
-            return result.ToActionResult();
-        }
-
-        return Ok(_mapper.Map<TrainerManagedPlanResult, TrainerManagedPlanDto>(result.Value));
-    }
-
-    [HttpPost("trainees/{traineeId}/plans/{planId}/delete")]
-    [ProducesResponseType(typeof(ResponseMessageDto), StatusCodes.Status200OK)]
-    public async Task<IActionResult> DeleteTraineePlan([FromRoute] string traineeId, [FromRoute] string planId, CancellationToken cancellationToken = default)
-    {
-        if (!Id<UserEntity>.TryParse(traineeId, out var parsedTraineeId))
-        {
-            return Result<Unit, AppError>.Failure(new InvalidTrainerRelationshipError(Messages.UserIdRequired)).ToActionResult();
-        }
-
-        if (!Id<PlanEntity>.TryParse(planId, out var parsedPlanId))
-        {
-            return Result<Unit, AppError>.Failure(new InvalidTrainerRelationshipError(Messages.FieldRequired)).ToActionResult();
-        }
-
-        var trainer = HttpContext.GetCurrentUser();
-        var result = await _trainerRelationshipService.DeleteTraineePlanAsync(trainer!, parsedTraineeId, parsedPlanId, cancellationToken);
-        if (result.IsFailure)
-        {
-            return result.ToActionResult();
-        }
-
-        return Ok(_mapper.Map<string, ResponseMessageDto>(Messages.Deleted));
-    }
-
-    [HttpPost("trainees/{traineeId}/plans/{planId}/assign")]
-    [ProducesResponseType(typeof(ResponseMessageDto), StatusCodes.Status200OK)]
-    public async Task<IActionResult> AssignTraineePlan([FromRoute] string traineeId, [FromRoute] string planId, CancellationToken cancellationToken = default)
-    {
-        if (!Id<UserEntity>.TryParse(traineeId, out var parsedTraineeId))
-        {
-            return Result<Unit, AppError>.Failure(new InvalidTrainerRelationshipError(Messages.UserIdRequired)).ToActionResult();
-        }
-
-        if (!Id<PlanEntity>.TryParse(planId, out var parsedPlanId))
-        {
-            return Result<Unit, AppError>.Failure(new InvalidTrainerRelationshipError(Messages.FieldRequired)).ToActionResult();
-        }
-
-        var trainer = HttpContext.GetCurrentUser();
-        var result = await _trainerRelationshipService.AssignTraineePlanAsync(trainer!, parsedTraineeId, parsedPlanId, cancellationToken);
-        if (result.IsFailure)
-        {
-            return result.ToActionResult();
-        }
-
-        return Ok(_mapper.Map<string, ResponseMessageDto>(Messages.Updated));
-    }
-
-    [HttpPost("trainees/{traineeId}/plans/unassign")]
-    [ProducesResponseType(typeof(ResponseMessageDto), StatusCodes.Status200OK)]
-    public async Task<IActionResult> UnassignTraineePlan([FromRoute] string traineeId, CancellationToken cancellationToken = default)
-    {
-        if (!Id<UserEntity>.TryParse(traineeId, out var parsedTraineeId))
-        {
-            return Result<Unit, AppError>.Failure(new InvalidTrainerRelationshipError(Messages.UserIdRequired)).ToActionResult();
-        }
-
-        var trainer = HttpContext.GetCurrentUser();
-        var result = await _trainerRelationshipService.UnassignTraineePlanAsync(trainer!, parsedTraineeId, cancellationToken);
-        if (result.IsFailure)
-        {
-            return result.ToActionResult();
-        }
-
-        return Ok(_mapper.Map<string, ResponseMessageDto>(Messages.Updated));
     }
 }
